Unify the output of the styles folder build
Do a funky thing with how we reference a global animation to avoid the
usage of `:global` in our mixins. This means that contents of the mixins
for use by consumers who don't use css-modules (/styles) and those that
do use css-modules (/esnext/styles) are identical.

This allows us to remove some manipulation in the build now.
Once we start shipping built instead of source scss files in the esnext
folder we would be able to remove esnext/styles entirely and tell
consumers that they can always use the api via `styles/_public-api.scss`.